requirements.txt
fastapi
uvicorn[standard]
sqlmodel>=0.0.8
sqlalchemy>=1.4
pydantic>=1.10
python-dotenv>=1.0.0
